Fix incorrect condition when skipping trivial casts
Summary:
PUNT() makes it possible to reach the CastVec opcode without having a guard. If
there was a value of unsupported type on the stack, such as LazyCls, we would
PUNT(), but then the next round would interpret the CastVec, not needing the
guard.

This is the rare case when bespoke's skipTrivialCast() can't make decision
statically and emits a branch to skip the trivial cast. Unfortunately the
condition was backwards and a non-matching type represents CheckType's
"then" branch.

So, fix this condition.

Reviewed By: ricklavoie

Differential Revision: D39797745

fbshipit-source-id: 60c71f317d534ed8ddf359b13f155a3ab7d4ab80