Fix treatment of IsTypeOp::Func and IsTypeOp::ClsMeth
Summary:
is_fun() and is_class_meth() return true for function pointers bound with
reified generics, so they need to support TRFunc and TRClsMeth.

The impact of this bug was quite limited, because of lack of inference logic
that would create standalone TRFunc and TRClsMeth types. However, this bug
caused a refcounting issue in test/slow/function_pointer/null_comparison.php,
where it inferred that `$x` in a cleanup block of is_reified() would be always
uncounted. This manifested as an assert when EnableArgsInBacktraces=0.

Reviewed By: ricklavoie

Differential Revision: D39803653

fbshipit-source-id: ebafe5e063dffbb1b471c2fe5f3ac711f6fb6ed5