Do not release text/declaredEncoding, while still used by EncodingMatch
Summary:
EncodingDetector::detect() constructs an EncodingMatch, which holds onto the
underlying detector, as the match is owned by it.

However, it did not hold onto text and declaredEncoding strings that are used,
but not owned, by the detector. So if the EncodingDetector was freed, or
setText()/setDeclaredEncoding() methods were called, these strings may have
been freed and the subsequent usage of the EncodingMatch would use a garbage.

Fix this by holding onto these strings alongside the detector.

Reviewed By: ricklavoie

Differential Revision: D39805277

fbshipit-source-id: 11106ec5e0f0866bc9523835f6561f1ca7d8b29a