&(&str, Decl)
Summary:
This diff does similar to (abandoned) D39786286, but in a more rust-idiomatic way.

------------------

Here's what we get when we parse a file:
```
pub struct ParsedFile<'a> {
    pub decls: Decls<'a>,
}

pub struct Decls<'a>(
    pub  arena_collections::list::List<'a, (&'a str, oxidized_by_ref::shallow_decl_defs::Decl<'a>)>,
);

pub enum Decl<'a> {
    Class(&'a ClassDecl<'a>),
    Fun(&'a FunDecl<'a>),
    Typedef(&'a TypedefDecl<'a>),
    Const(&'a ConstDecl<'a>),
    Module(&'a ModuleDecl<'a>),
}
```
When we iterate decls, we get back Decl as a value-type. (after all, it's only a tiny enum).
```
pub fn iter(&self) -> impl Iterator<Item = (&'a str, Decl<'a>)>
```

***BUT*** this conflicts with arena.add_root:
```
pub fn add_root<'a, T: ToOcamlRep + ?Sized>(&'a self, value: &'a T) -> Value<'a> {
```
In particular, if you have local variable of value-type `decl:Decl` and call `arena.add_root(&decl)` then the returned value has a lifetime 'a which is the shortest of (1) the arena lifetime, (2) the lifetime of the local variable `decl`.

That's not very good! The whole point of add_root is that it copies the thing out of the "value" argument you passed.

With this diff, I changed the iterator:
```
pub fn iter(&self) -> impl Iterator<Item = &'a (&'a str, Decl<'a>)>
```
That way, anyone who uses .iter() will be able to get an &'a Decl with the same lifetime as the bumpalo.

-------------

I left the IntoIterator the same:

```
impl<'a> IntoIterator for Decls<'a> {
    type Item = (&'a str, Decl<'a>);
    type IntoIter = std::iter::Copied<arena_collections::list::Iter<'a, (&'a str, Decl<'a>)>>;
```

That feels kind of weird to me -- an IntoIterator is supposed to take ownership of the underlying collection, but you can't take ownership of something that lives in an arena. But anyway.

Reviewed By: edwinsmith

Differential Revision: D39783977

fbshipit-source-id: 28c4787ac47af0a1463c1174b2044ebddea615b1