Build hhvm_clang with libstdc++ instead of libc++ (#9213)
Summary:
This PR builds `hhvm_clang` nix package with libstdc++ instead of libc++.

Note that this PR also switch clang to version 14 instead of version 12, because it is a closer version to the internal clang version 15, given that nixpkgs does not provide clang 15 right now. See https://github.com/NixOS/nixpkgs/issues/191132

Pull Request resolved: https://github.com/facebook/hhvm/pull/9213

Test Plan: This PR should not affect `hhvm` nix package, which is built with gcc. The CI job for `hhvm_clang` is expected to fail because we have not ported HHVM OSS to Clang.

Reviewed By: alexeyt

Differential Revision: D39673905

Pulled By: Atry

fbshipit-source-id: 28906d5dbd243ecbdea90ccab9a8d611fa2a8f99