Fix test/slow/ext_apc/apc_extend_ttl.php flakiness
Summary:
APC stores timestamps in integer number of seconds. This means if a TTL of N is
requested, the value will expire between N-1 and N seconds.

apc_extend_ttl.php had a habit of setting/extending TTL to 3 seconds and then
sleeping for 2 seconds. Once a while the timing was unfortunate and the value
expired right after the sleep, before we managed to obtain it.

Change timings so that there is always at least 1 second of margin for a delay
caused by scheduling/jitting/etc. Note that a 3 second sleep is guaranteed to
expire a value with TTL of 3 seconds, no margin needed in that direction.

Reviewed By: binliu19

Differential Revision: D39840146

fbshipit-source-id: c3da0a868274cc48964ca68d8bae15d2cec0f1de