Back off to dynamic method invocation in dynamic mode
Summary:
When checking method calls in dynamic mode, we should support subsumption to `dynamic` if an attempt at checking a static call fails. For example, if we have a call through `$x : J` to a method `foo` that doesn't exist on `J`, then we should proceed to check the call at `$x : dynamic`.

The need for this arises when types are simplified e.g. `~C & J` is simplified to `J` if `C` implements `J` (and both are SDT), because under dynamic-aware subtyping, `~C` is equivalent to `dynamic`, and `dynamic & J` is equivalent to `J`.

Reviewed By: vassilmladenov

Differential Revision: D39845935

fbshipit-source-id: 5a8e0b3915d7f8441b06c713671a761c087d1b25