Explicitly define a user-provided move constructor for OmniClient
Summary: For reasons that are unclear to me, when compiling the OmniClient with non-internal toolchains `std::atomic` no longer has a defined constructor. I'm not smart enough to be able to parse the standard well enough to explain why this is happening, but I do know how to fix it! This commit adds an extremely naive explicit move constructor to OmniClient which should (hopefully) resolve the issue.

Reviewed By: splbio

Differential Revision: D39859770

fbshipit-source-id: 0700ee92e493357995000679acd11118878443ba