Only bump limit when under threshold and seeing errors
Summary:
Today we bump limit when we are under threshold of our cpu target and also seen active requests / qps in last interval that is close to our limit.
This is a pretty lossy approach as estimating QPS/active requests in short interval is not accurate. I suspect this is causing some poor maintenance of target under overload.

This change adds a new configuration where we only bump our limit if we are under target and actually saw load shedding since last interval. Bonus points is that this is simple to configure and should be the default.

Reviewed By: sethdelliott

Differential Revision: D39853263

fbshipit-source-id: 1fbec013bf632f73c72ac81c33ae4302d24f635f