Wrap the given DeclProvider with SelfProvider
Summary:
This stack intends to have HackC able to use decls from the current file when compiling. This is most useful for CLI mode or script mode where the entry file being compiled does not show up in the autoloader because it exists in a directory that is not indexed by the autoloader.

In the previous diffs, I have created the SelfProvider DeclProvider. In this diff, I wrap the given DeclProvider in the SelfProvider.

Note: I'm not convinced this is the best design, so happy to hear suggestions.

Reviewed By: edwinsmith

Differential Revision: D39331187

fbshipit-source-id: 64c26de5f1ed0b1b2a7728732e4294a9c06f9c18