Create a new DeclProvider that provides decls from the current file
Summary:
This stack intends to have HackC able to use decls from the current file when compiling. This is most useful for CLI mode or script mode where the entry file being compiled does not show up in the autoloader because it exists in a directory that is not indexed by the autoloader.

This diff introduces a new DeclProvider called SelfProvider. It takes an optional decl provider that it queries as a backup if the self decl provider doesn't have the decl.

Reviewed By: edwinsmith

Differential Revision: D39323439

fbshipit-source-id: 848906f9cb0a92ed8497ed838260eb9084e37924