Skip locale-using tests if locales used are not available
Summary: A number of tests fail in environments where we run cmake builds, apparently because the locales we use in those tests are not available. Make these tests all declare their dependency on the locales they use so that they can be skipped when the locales are not present on the system running the test.

Differential Revision: D39861973

fbshipit-source-id: 41a3d78572d42b5b707129108a488232719a0218