Run all tests (with some exclusions) in github actions (#9230)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/9230

Unfortunately run.php doesn't support multiple exclude files, so this diff copies the file with the smaller set of excludes used for local cmake testing and expands it. For now, if we want to mutate the list of excluded tests we will need to touch both files (but my hope is to get them to be identical and much smaller soon, which will allow us to have only one exclude file in the future).

Differential Revision: D39861976

fbshipit-source-id: baa4f55a24f69c8aa2a8279c26cc4571fa3e8e8d