Use intersection types for partial enforcement
Summary:
Our current pessimisation strategy is to add a like type to any return type that is not fully enforced by HHVM. When subtyping in Hack agrees (`t <: u`) with "partial" enforcement in HHVM (e.g. `t` is enforced at type `u`), we end up with situations such as the following:
```
interface I {
  public function get():J;
}
class C<T as J> {
  public function get(): T { ... }
}
```
This should get pessimised to
```
interface I {
  public function get():J;
}
class C<T as J> {
  public function get(): ~T { ... }
}
```
but because `~T` is not a subtype of `J`, we end up *poisoning* the hierarchy i.e. adding a like-type to `I::get`.

Instead of this non-local poisoning, we can make use of intersection types to express that `T` is enforced at its bound:
```
interface I {
  public function get():J;
}
class C<T as J> {
  public function get(): (~T & J) { ... }
}
```
Of course, this is now relying on HHVM enforcing the first component of the intersection (i.e. the original type), which right now, it doesn't do, but could easily be modified to do so in future. (Or we express all of the above in implicit pessimisation.)

To implement the above, we define a function `get_enforced_type` that maps a decl type `t` to a decl type `u` known to be enforced by HHVM. We require `t <: u`. Sometimes, `u` is not denotable e.g. the enforced type of `MyVec<int>` is `MyVec<_>` where `_` is a wildcard (existentially quantified). We can't intersect with this directly. So instead, we have the following more complex transformation, here illustrated by interfaces but the same applies to any kind of inheritance:
```
interface I1 {
  public function foo(): t1;
}
...
interface In {
  public function foo(): tn;
}
class C implements I1, ..., In {
  public function foo(): t;
}
```
Suppose that all of `t1` to `tn` are fully enforced, and that `t` is partially enforced at type `u`. If `u <: ti` for all the `ti`s, we pessimise `C::foo` as follows:
```
  public function foo(): (~t & t1 & ... & tn);
```
For any `ti` that is fully enforced but for which it's not the case that `u <: ti`, we must "poison" that ancestor i.e. add a like type. The main example of this is type constants, which are not enforced at all. For example:
```
class C {
  const type TC = int;
}
interface I1 {
  public function foo(): int;
}
class C implements I1 {
  public function foo(): C::TC;
}
```
In this situation, we poison the hierarchy, ending up with
```
interface I1 {
  public function foo(): ~int;
}
class C implements I1 {
  public function foo(): ~C::TC;
}
```
Note that in future we might move to a place where `C::TC` *is* enforced, or we might enforce all components of intersections, in which case `(~C::TC & int)` would be a suitable pessimisation of `C::TC`. Either of these is a behaviour change for HHVM.

Reviewed By: vassilmladenov

Differential Revision: D38832563

fbshipit-source-id: 65ad6b26f5e47500560d008b54728a37ed3e6be1