Fix arguments in editline tests that only run in cmake builds
Summary:
We failed to update these unit tests when we banned argument coercion for builtin arguments (because we never run these tests internally). Try to fix these tests by replacing values of the wrong type with the values that they would have been coerced to when that was still allowed.

Also fix one test expectation that was not adjusted to deal with hack array migration.

Reviewed By: Atry

Differential Revision: D39861971

fbshipit-source-id: e6b9f750182cba77dff033f1014c6632d6229455