RFC: `BuiltinSymbolMap`
Summary:
"Systemlib" is the set of PHP files that are shipped with HHVM by being directly embedded in the binary. They're compiled and loaded independently of other PHP code and not included in autoloading. This hides them from the current architecture for decl driven bytecode compilation, which is autoloader based.

Given that we unconditionally compile these files on HHVM startup, it makes sense to me to unconditionally parse their decls and store them, allowing `HhvmDeclProvider` to access these blobs if we fail to find a symbol in the autoloader. Some notes:

* Systemlib uses the *default* runtime options when being compiled, I've done the same here
* While the decls in systemlib are now exposed to other files, systemlib itself still cannot be compiled with decls, we fail to give it a provider.

Reviewed By: paulbiss

Differential Revision: D38809048

fbshipit-source-id: c1250d1387ede938f12f4589e683f7f8245f5ee5