tighten up hh24_types dependencies 2/2
Summary:
This diff makes hh24_types no longer depend upon "oxidized".

It was using it only for an ugly trick. This is how it worked...
* `names` knew the oxidized::NameType for a symbol, since that's how it's stored in naming-sqlite
* `typing_deps_hash` knew how to turn a NameType into a DepType
* `hh24_types` used that DepType to generate a hash

We need hh24_types::ToplevelSymbolHash to depend upon typing_deps_hash because we need the ToplevelSymbolHash to be bitwise identical to the DepHash for all top-level symbols. So they need to be in the same module.

But that conversion from NameType to DepType doesn't belong in hh24_types nor in typing_deps_hash. It's really related just to the `names` module.

Reviewed By: zhyty

Differential Revision: D39872075

fbshipit-source-id: 09fbb734eaa0faf2377f1f66277002476e1b2de7