Intersect enforced enum and newtype returns
Summary:
Update decl enforceability to compute what part of a type is enforced, and use that in the return.

NB, this is using the actual enforced type in the return, but this is too permissive because it throws away the  abstraction that the `as` bound (if present) is trying to provide. I will fix that in a subsequent diff.

Reviewed By: vassilmladenov

Differential Revision: D39814945

fbshipit-source-id: 6b3183b441f84b2b4674f502f3e2343d09c4d08d