Calculate fanouts due to changes in modularity more intelligently
Summary:
This diff optimizes the fanout algorithm for modules a bit better than always creating a Major change when a class's module changes.

Specifically if a class is public, when a class's module changes, we can calculate the fanout of its members normally with one exception. If the member's visibility is internal and the module has changed at all, we should consider that the member is modified (since the module affects the visibility of the internal keyword).

Therefore, if you have a fully public class and you add or modify its module, no fanout should occur.

In the future, once we have module references, we will need to be even smarter about this, since "fully public" classes changing modules also changes their visibility. But we can worry about that once we're ready to ship V2.

Reviewed By: panopticoncentral

Differential Revision: D39893840

fbshipit-source-id: 2e2d96e47bee33cb85b93933a1e2c6f29f353c7c