Support implicit upcast for SDT function application
Summary:
SDT function application doesn't support implicit upcast to dynamic. (See test.)

Why? Because the test for `dynamic` when inserting implicit upcasts is very syntactic, just looking for `Tdynamic`. Probably we should improve this for future, but for now an easy fix is to avoid unioning with dynamic when constructing a like-type of dynamic itself.

Differential Revision: D39966512

fbshipit-source-id: 75ec74e049cfb9ff747ec2154fb526730bcf372b