Optimise conjunctive forwarding
Summary:
Definite static keys are determined by conjunctive forwarding that is if the key exists in all the nodes from the incoming edges to a node, then the node also has that key.

This operation is expensive as it required iterating over all accumulated static key constraints (barring the semi-naïve optimisation). To further speed it up we now maintain a hash table of the static key constraints which replaces repeated traversals with constant time lookup (and hash table maintenance cost).

Differential Revision: D39767123

fbshipit-source-id: 93d986236f7d8ff28cbededa09c8fe362faaa7f2