Coarse collection support
Summary:
This diff adds simple-minded support for collections.

With the extra support, we can discover more shapes that are nested in collections.

However, the handling is heavily imprecise. We should have a structured entity that distinguishes between the collection and the elements within the collection. However, adding this support requires adding a new entity kind and solver changes. Instead, we conflate the entity for the collection with the elements in the collection. The postive consequence is that we can handle the following code just fine:

```
$v = vec[dict['a' => 42]];
inspect($v[0]); // yields shape('a' => int)
```

However, we also find spurious results such as
```
$d = dict['b' => dict['a' => 42]];
inspect($d); // shape('b' => dict<string, int>, 'b' => int);
```

Interestingly, the diff also reduces false positives in local mode because we can better detect nested dictionaries that escape local definitions, e.g., in
```
function f(): dict<string, mixed> {
  return dict['b' => dict['a' => 42]];
}
```
we no longer report an allocation for the nested dictionary as we see that it flows into the return.

The diff
- extends existing collection indexing
- extends existing collection overwrite
- extends existing collection literal
- adds new collection literals (`keyset`, `Set`, `vec`, `Vector`, `Map`, etc.)
- adds collection append (for `vec` and `Vector`)

Differential Revision: D39545143

fbshipit-source-id: c05dee93852acccb1d9faa5222c2dd0055ce2e31