Relax what hints should be kept track of
Summary:
We decide to keep track of parameter or return hints based on the type. This was quite strict and looked for a `dict` or an `Awaitable` of a `dict`.

This diff relaxes it by checking if the type is disjoint from dict and if not we keep track of it.

This helps reducing some common cases of false positives where there is a dict that escapes its scope, but there is no entity generated for the return/parameter hint which would be tainted with dynamic access to invalidate the result.

The current treatment, however, doesn't account for awaitables of `dict`s. This is tracked in T132035432.

Differential Revision: D39545145

fbshipit-source-id: ae0cf0a5654696fb969ad0b665832c830cfa9a7e