Error if a class does not use a trait that require class it
Summary:
If a trait `T` `require class C` but class `C` does not use `T`, then the upper bound for `this` is `C & T`.  Since `C` and `T` have an empty intersection, when resolving element search (eg `$this->foo()`), the type of `this` is simplified into `nothing` and everything typechecks.

Although this is not a typehole, the lack of any error is likely to be confusing to the programmer.  It also confuses e.g. code completion.

This diff raises an error whenever a trait `require class` a class that does not use it.

Reviewed By: vsiles

Differential Revision: D39851269

fbshipit-source-id: e51581dbd6f37eda6c51ec716f45c909fd659ca2