Add the boxed_ptr<T> class
Summary:
boxed_ptr<T> provides copy-on-write behavior for a value of type T. This
wrapper can only be used for copyable types. Trivial types and types such as
st::string that are already boxed are also disabled.

This class is intended to replace boxed_value_ptr<T> once code generation is
able to recognize when not to use boxed_ptr<T>.

Reviewed By: Alfus

Differential Revision: D39805636

fbshipit-source-id: 162a968a3ba97918026f89f98291ae1917a285b7