Parse more options
Summary:
We'd like to support invoking the type checker from inside a rust_binary
that parsed .hhconfig itself using hh_config::HhConfig. And for that to work
properly, HhConfig needs to parse more of the available options.

In addition, any config line that wasn't parsed is saved in an `unknown` field.
This is currently unused, but it shows up in Debug output as a quick sanity check
to make sure we didn't accidentally forget or misspesl an option.

Differential Revision: D39802665

fbshipit-source-id: 72839a6688494c22c2c5c8c9230cb72eb7e5a0c4