Do not crash when suspend event hook finishes the child awaitable
Summary:
When a previously suspended async function is suspending again because of an
await of an unfinished child, we may enter a suspend hook, which may invoke
a userland profiler. If that profiler causes the child awaitable to finish, the
suspend logic will encounter an unexpected state and crash.

Let's throw an exception, which the suspend logic is still capable of handling
at this state. This is not optimal, but better than crashing.

Ideally we would be capable of continuing the execution. The issue is that this
would either cause an additional code bloat per every await, or additional
runtime checks on the hot code path.

Reviewed By: ricklavoie

Differential Revision: D40003779

fbshipit-source-id: 8c4e735bdba4adbc17be9aba78b7d40122b273c1