Fix typing for dynamic class member access
Summary:
# Background

Currently, there is a typehole around the following bit of code:

```
$x::$y()
```

HHVM treats this as a dynamic method call to the `$x::value_found_in_y()` method, where `value_found_in_y` is the string name contained in local `$y`. Hack treats this as if it were invoking the callable stored in the static property `y` on the class at `$x`, which is more properly expressed as `($x::$y)()`.

This confusion mainly materializes around the fact that we carry around the `is_prop_or_method` field on the `Class_get` node but never seem to do anything important with it.

# This change

We ensure that `ClassGetExpr::CGstring` and `ClassGetExpr::CGexpr` flavors are interpreted consistently in the typing code. Specifically, we assume `CGstring` is a static property whose name is statically known, and all others compile to `CGexpr`.

Differential Revision: D38969293

fbshipit-source-id: 63bb1ab45e73898c7aeba7b06009505ea14737df