Introduce naming error for dynamic call to class method
Summary:
```
Foo::{'ba'."r"}();
```
Attempts to call the static method `bar` on class `Foo`. This is a form of dynamic method invocation that this diff bans.

This allows us to be a bit more precise in the error messaging and the range on which we report the error. We also set up for the next diffs in which we fix recognition of `Bar::$y()` as a static call to the method on `Bar` whose name is in the local variable `$y`, bringing Hack in line with the runtime behavior.

Differential Revision: D39899450

fbshipit-source-id: c7a897d6e16cf660ba16a2bc3b20eb49d35ba310