Fix lowering of static calls with variable method name
Summary:
# Background

Currently, there is a typehole around the following bit of code:

```
$x::$y()
```

HHVM treats this as a dynamic method call to the `$x::value_found_in_y()` method, where `value_found_in_y` is the string name contained in local `$y`. Hack treats this as if it were invoking the callable stored in the static property `y` on the class at `$x`, which is more properly expressed as `($x::$y)()`.

This confusion mainly materializes around the fact that we carry around the `is_prop_or_method` field on the `Class_get` node but never seem to do anything important with it.

The previous diffs in the stack updated the typing modules to consistently treat `ClassGetExpr::CGstring` as statically-known class member access (only applicable to properties) and `ClassGetExpr::CGexpr` as holding any dynamic class member name requiring evaluation as an expression.

# This change

Here we update the lowerer to account for context in determining whether a `::` (scope resolution) expression dynamically refers to some class member (=> `CGexpr`) or refers to a static property (=> `CGstring`). In conjunction with the previous change it fixes the typehole.

Curiously, HackC doesn't require any changes because it was already accounting for this intent when it compiled to `CGstring` instead of `CGexpr`. These blocks should now be unreachable:
https://www.internalfb.com/code/fbsource/[a9bac2fd5ac2def55491d6cbbd613f7f751fa4cc]/fbcode/hphp/hack/src/hackc/compile/closure_convert.rs?lines=1094-1096
https://www.internalfb.com/code/fbsource/[5598f544df4f8af75e315a693ca5df8ba22d2259]/fbcode/hphp/hack/src/hackc/emitter/emit_expression.rs?lines=2171-2173
https://www.internalfb.com/code/fbsource/[5598f544df4f8af75e315a693ca5df8ba22d2259]/fbcode/hphp/hack/src/hackc/hhbc/decl_vars.rs?lines=64-70

Differential Revision: D39899449

fbshipit-source-id: cef67a3d98371111839a8525817341c1ee7a2eb8