Fix mutable iteration for runtime containers
Summary:
Consider a function

  void foo(const Iterable::value_type& bar)

The gtest suite has such a function, which it calls from:
  Iterable::const_iterator itr = iterable.begin();
  ...
  foo(*itr);

The problem is that `*itr` now return `Iterable::const_iterator::reference`, which is always a `ConstRef`. So to support this, we need to be able to implicitly convert from a `ConstRef` -> `const Ref&`. Luckily c++ supports this natively when `ConstRef` is passed by `&&` **and** its largely nonsensical to pass `ConstRef` by `&&` to the constructor of `Ref` in any other context.

Differential Revision: D39858469

fbshipit-source-id: c077786c516f18fd69f13d0a6e39eeaabf660a83