delete oxidized::relative_path::to_absolute()
Summary:
```
    pub fn to_absolute(&self) -> PathBuf {
        let prefix_map = PrefixPathMap::default();
        let prefix = self.prefix.to_path(&prefix_map);
        prefix.join(&self.path)
    }
```
This function does not do what it says on the tin. It doesn't convert a relative_path into absolute. Instead it merely returns the suffix. (standalone demo: D40036559)

So I'm deleting it. If someone wants to see the suffix, they should call the standard path() function.

How is this done elsewhere?

Separately, oxidized_by_ref::to_absolute takes a structure so it knows the path of each prefix and so can properly do to_absolute.

Separately, in ocaml, Relative_path.to_absolute relies on ambient global state (which stores the absolute path).

Differential Revision: D40037035

fbshipit-source-id: 2b998d56ec8b6c4ac674de88b456ebb1994175db