Stop inconsistently using the TypedNum alias for TypedValue
Summary: We have an alias for TypedValue called TypedNum, which is supposed to mean the TV contains an int or double (but this is by convention only, there is no enforcement or behavior change). This was only applied in a handful of places, and the ony ones left were tvSub and tvMul from tv-arith.{h,cpp}, a tiny fraction of the places where we have such TypedValues. At this point it's more confusing than helpful, so delete the alias and just use the TypedValue type.

Reviewed By: paulbiss

Differential Revision: D40093634

fbshipit-source-id: a7ec457f5f40dfe464ecd1ed54af4430ec426308