Intersection <: Like
Summary: It's common under Sound Dynamic, especially with `--pessimise-builtins`, do require a subtype test of the form `t <: ~#1` where `#1` is a type variable. If the left hand side is an intersection, our current heuristic is to break apart the intersection. This often gives us the wrong answer: we would rather break apart the union, and end up with `t <: #1`.

Differential Revision: D40094780

fbshipit-source-id: b0a74e299175894c8d9fb7525707a6a9f3def203