Move code around to delete more C++ declarations of native funcs
Summary:
We're relying on declarations of some native builtins in extension headers because we define them earlier in their extension .cpp file than the other native builtins they call, or because we rely on the declaration to provide default values of some optional arguments.

If we move these definitions lower and/or provide all the arguments explicitly we can delete more of these C++ declarations.

In ext_datetime.cpp we can also replace a call to `f_time` (which is implemented as a call to `time(0)`) with a call to the underlying implementation (and stop declaring the `time` native builtin in C++).

Reviewed By: ricklavoie

Differential Revision: D40079994

fbshipit-source-id: a1f4af4b9d12d29f234960b6421a30096a4e3673