Move moduleInits to the bottom of their cpp files
Summary:
Some extensions have moduleInit functions defined near the top of their extension cpp file, which forces them to rely on the forward declaration of all the native functions/methods in the extension .h file in order for these functions to be visible and for the extension to compile.

Move them to the bottom of their cpp files to remove this dependency.

Reviewed By: ricklavoie

Differential Revision: D40079991

fbshipit-source-id: f7314e2720b1a8ecceb6c21f0981aa34323469c4