store checksum in naming-sqlite
Summary:
For the re-architecture, we will use a "checksum", something that identifies the current snapshot of all decl content in the entire repository -- if any symbol's decl changes, or if a symbol gets added or deleted, or if a symbol moves into a different file, then the checksum will change.

The checksum is computed from the DECL_HASH which is already stored in the naming-table (computed in rust from hash(oxidized_by_ref::shallow_decl_defs:decl), and HASH which is  already stored in the naming-table (it's the primary key for the NAMING_SYMBOLS table, and also the same as 64bit depgraph hash for top level symbols). The checksum calculation is simply
```
checksum = checksum XOR hash(path,symbol_hash,decl_hash)
```

I added checksum into two places: (1) when hh_server produces a naming-sql from scratch, where it just does the XOR operation for every symbol, (2) when hh_server updates a naming-sql based on a delta, where is does XOR to remove all old checksum contributions from files modified in the delta, then XOR to add all new checksum contributes from files modified in the delta.

Reviewed By: bobrenjc93

Differential Revision: D40146748

fbshipit-source-id: b651e9267844553959b2219db9df3647e8b02ee8