show expected/actual output mismatch
Summary:
This test code used python's jedi library `assertCountEqual` function. Its name is tricky. It asserts that if you sorted the two lists then their sorted versions would be identical. Another way of saying is that the frequency count of each line is identical in the two.

Not surprisingly this means that output messages are hard to read when they're different! The output messages say "LHS had 1 copies of this string but RHS had 2 copies" or similar.

I've changed it to also show just a straightforward statement of what was the output, and what was expected.

Reviewed By: bobrenjc93

Differential Revision: D40158850

fbshipit-source-id: 548cb04e372c96bb6a291714d5fb14b80f6b1770