Add source info to schema.thrift
Summary:
Some elements of the schema are already context-dependent (e.g. structuredAnnotations docblock mentions different filters in schema and AST contexts). Adding a small amount of source info to schema.thrift only in AST context makes the representation usable directly as the AST, which will make keeping the two representations up-to-date trivial and increases the reusability of the schema generation utilities for AST generation.

These updated definitions can then be used to write a universal compiler frontend for feeding to thrift compiler backends written in other languages, and to source indexers.

Reviewed By: vitaut

Differential Revision: D40031398

fbshipit-source-id: 00f24ac7c6de0b346d92361a73a33376e131bfd0