Enable checking only global writes or reads
Summary: The global access checker is updated to accommodate with two new arguments "`--disable-global-access-check-on-write`" and "`--disable-global-access-check-on-read`" introduced in D40066119 (https://github.com/facebook/hhvm/commit/a78f1fdf16275be37fd65b94485f67487a68e152), such that the error regarding global writes/reads is raised only if it is NOT disabled. If both are disabled, then the checker will not run at all.

Reviewed By: yuxuanchen1997

Differential Revision: D40078957

fbshipit-source-id: 0e8d7dfe526a6aef9b9b2727d2ff55734d4a566c